added laptop matrices, cleaned up code
